fix: use architecture-aware RHEL preference in snapshot fixtures
The rhel_vm_for_snapshot and rhel_vm_for_snapshot_with_content fixtures
hardcode RHEL10_PREFERENCE which enables SecureBoot. On s390x, SecureBoot
is unsupported (no EFI OVMF ROMs), causing VM boot failures.

Replace hardcoded RHEL10 preference and data source with dynamic lookup
via py_config["latest_instance_type_rhel_os_dict"], which resolves to
RHEL9 on s390x (no SecureBoot) and RHEL10 on amd64/arm64.

Add latest_rhel_data_source_scope_session fixture to tests/conftest.py
that dynamically selects the data source based on architecture.
